Simplify interface event functions.
