+++ eduvpn/interface/__init__.py
@@ -1,3 +1,4 @@
 # Avoid import order issues.
 from . import event  # noqa: F401
+from . import transition  # noqa: F401
 from . import state  # noqa: F401
